+++ CHANGELOG.md
@@ -9,6 +9,7 @@ backwards compatibility.
 * **Breaking** Remove logging abstractions from the crate.
 * **Breaking** Bump STM32H7 HAL to version 0.14.
 * Improve style and documentation of examples.
+* Introduce RTIC example of audio processing.
 
 ## 0.7.3